Fix a bug in 31 bit arithmetic, leading to failing conversion tests.
On 64 bits architectures, integers could have some of their 32 msb set to 1
internally in the VM. When read back to a Coq term, this was not observable. But
an equality test would fail. From the user point of view, the symptom was that
vm_compute; reflexivity would succeed but the subsequent Qed would fail.

Bug reported by Tahina Ramananandro.

+++ kernel/byterun/coq_interp.c
@@ -1241,7 +1241,7 @@ value coq_interprete
         if (shiftby > 31) {
           if (shiftby < 62) {
             sp++;
-	    accu = (value)((((*sp++)^1) << (shiftby - 31)) | 1);
+	    accu = (value)(((((uint32_t)*sp++)^1) << (shiftby - 31)) | 1);
           }
           else {
             sp+=2;
@@ -1250,7 +1250,7 @@ value coq_interprete
 	}
         else{
         /* *sp = 2*x+1 --> accu = 2^(shiftby+1)*x */
-        accu = (value)(((*sp++)^1) << shiftby);
+	  accu = (value)((((uint32_t)*sp++)^1) << shiftby);
         /* accu = 2^(shiftby+1)*x --> 2^(shifby+1)*x+2*y/2^(31-shiftby)+1 */
         accu = (value)((accu | (((uint32)(*sp++)) >> (31-shiftby)))|1);
 	}
